Rename namespace, urls and databases to match new repo name
