+++ .github/workflows/ci.yml
@@ -20,11 +20,11 @@ jobs:
       php_extensions: memcached, mysql, bz2, gd, igbinary, imagick, imap, mbstring, bcmath, zip, intl
       COMPOSE_INTERACTIVE_NO_CLI: 1
       WP_ROOT_FOLDER: ${{ github.workspace }}/tests/wordpress
-      WP_DOMAIN: "square1-acf-field-models.tribe"
+      WP_DOMAIN: "square1-field-models.tribe"
       WP_ADMIN_PATH: "/wp-admin"
       WP_ADMIN_USERNAME: "admin"
       WP_ADMIN_PASSWORD: "password"
-      WP_URL: "http://square1-acf-field-models.tribe"
+      WP_URL: "http://square1-field-models.tribe"
       TEST_DB_HOST: 127.0.0.1
       TEST_DB_NAME: wordpress
       TEST_DB_USER: root
