+++ src/Collections/User_Collection.php
@@ -1,9 +1,9 @@
 <?php declare( strict_types=1 );
 
-namespace Tribe\Libs\Acf_Field_Models\Collections;
+namespace Tribe\Libs\Field_Models\Collections;
 
 use Spatie\DataTransferObject\DataTransferObjectCollection;
-use Tribe\Libs\Acf_Field_Models\Models\User;
+use Tribe\Libs\Field_Models\Models\User;
 
 class User_Collection extends DataTransferObjectCollection {
