+++ src/Models/Image.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Tribe\Libs\Acf_Field_Models\Models;
+namespace Tribe\Libs\Field_Models\Models;
 
 class Image extends File {
