+++ tests/.env-dist
@@ -1,11 +1,11 @@
 WP_ROOT_FOLDER="tests/wordpress"
-WP_DOMAIN="square1-acf-field-models.tribe"
-WP_URL="http://square1-acf-field-models.tribe"
+WP_DOMAIN="square1-field-models.tribe"
+WP_URL="http://square1-field-models.tribe"
 WP_ADMIN_PATH="/wp-admin"
 WP_ADMIN_USERNAME="admin"
 WP_ADMIN_PASSWORD="password"
 TEST_DB_HOST="mysql.tribe"
-TEST_DB_NAME="tribe_acf_field_models_tests"
+TEST_DB_NAME="tribe_field_models_tests"
 TEST_DB_USER="root"
 TEST_DB_PASSWORD="password"
 TABLE_PREFIX="tribe_"
