+++ tests/integration/Tribe/Libs/Field_Models/FieldModelTest.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Tribe\Libs\Acf_Field_Models;
+namespace Tribe\Libs\Field_Models;
 
 use Tribe\Libs\Tests\Test_Case;
